refactor(wikidata-client): extract generic sparql-client crate
Split the SPARQL-over-HTTP machinery out of wikidata-client into a new,
endpoint-agnostic `sparql-client` crate, leaving wikidata-client as a thin
biodiversity layer (GBIF cross-walks, taxon images) on top.

This separates the two concerns the crate had conflated — a reusable
generic SPARQL client vs. Wikidata/GBIF-specific helpers — and puts the
seam where a future extraction to a dedicated repo would cut.

- New crate `sparql-client`: SparqlClient (endpoint arg + getter),
  SparqlValue/SparqlBinding, Error (is_throttled/is_timeout),
  escape_literal, sparql_query/sparql_ask. Deps: reqwest, serde.
- wikidata-client now wraps a SparqlClient pinned to the Wikidata
  endpoint; keeps the GBIF helpers, is_property_id, and to_thumbnail_url,
  and re-exports Error/SparqlBinding/SparqlValue. Drops the direct
  reqwest + serde deps.
- WikidataClient public API is unchanged for callers (new,
  with_user_agent, sparql_query, and the four helpers); the
  externally-unused with_endpoint/try_with_user_agent/sparql_ask now live
  on SparqlClient. observing-appview and observing-resolve-taxa compile
  untouched.
